assigned random to cp in the seeds database

+++ db/migrate/20160805091837_create_pokemons.rb
@@ -2,7 +2,7 @@ class CreatePokemons < ActiveRecord::Migration[5.0]
   def change
     create_table :pokemons do |t|
       t.string :name
-      t.integer :cp
+      t.float :cp
       t.string :image_url
 
       t.timestamps
